fix(run): update aws-cdk
This updates the base image and the python lib versions.

I needed to do this because I encountered this bug in aws-cdk:

https://github.com/aws/aws-cdk/issues/14738

Just doing the workaround inside a Dockerfile based on
`contino/aws-cdk` did not work for me.  But rebuilding the image
itself did.

Related PR: https://github.com/contino/docker-aws-cdk/pull/6

**Don't merge before the above PR is merged and the new image released**